requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-tinymce==3.4.0
Pillow==9.1.1
qrcode==7.3.1
sqlparse==0.4.2
